Added data validation

+++ Controllers/AnimeController.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 
 
-[Route("api/[controller]")]
+[Route("api/anime")]
 [ApiController]
 public class AnimeController : ControllerBase
 {
@@ -16,6 +16,9 @@ public class AnimeController : ControllerBase
     [HttpGet]
     public async Task<IActionResult> GetAllAnimes()
     {
+        if(!ModelState.IsValid)
+            return BadRequest(ModelState);
+
         var animes = await _animeRepo.GetAllAnimesAsync();
         
         var animeDto = animes.Select(s => s.ToAnimeDto());
@@ -23,9 +26,12 @@ public class AnimeController : ControllerBase
         return Ok(animeDto);
     }
 
-    [HttpGet("{animeId}")]
+    [HttpGet("{animeId:int}")]
     public async Task<IActionResult> GetAnimeById([FromRoute] int animeId)
     {
+        if(!ModelState.IsValid)
+            return BadRequest(ModelState);
+        
         var anime = await _animeRepo.GetAnimeByIdAsync(animeId);
 
         if(anime == null)
@@ -36,9 +42,12 @@ public class AnimeController : ControllerBase
         return Ok(anime.ToAnimeDto());
     }
 
-    [HttpPost("{genreId}")]
+    [HttpPost("{genreId:int}")]
     public async Task<IActionResult> CreateAnime([FromRoute] int genreId, CreateAnimeRequestDto animeDto)
     {
+        if(!ModelState.IsValid)
+            return BadRequest(ModelState);
+        
         if(!await _genreRepo.GenreExists(genreId))
         {
             return BadRequest("Genre does not exist");
@@ -52,9 +61,12 @@ public class AnimeController : ControllerBase
     }
 
     [HttpPut]
-    [Route("{animeId}")]
+    [Route("{animeId:int}")]
     public async Task<IActionResult> UpdateAnime([FromRoute] int animeId, [FromBody] UpdateAnimeRequestDto updateAnimeDto)
     {
+        if(!ModelState.IsValid)
+            return BadRequest(ModelState);
+        
         var anime = await _animeRepo.UpdateAnimeAsync(animeId, updateAnimeDto.ToAnimeFromUpdateDTO());
 
         if(anime == null)
@@ -66,9 +78,12 @@ public class AnimeController : ControllerBase
     }
 
     [HttpDelete]
-    [Route("{animeId}")]
+    [Route("{animeId:int}")]
     public async Task<IActionResult> DeleteAnime([FromRoute] int animeId)
     {
+        if(!ModelState.IsValid)
+            return BadRequest(ModelState);
+        
         var animeModel = await _animeRepo.DeleteAnimeAsync(animeId);
 
         if(animeModel == null)
